refactor(analyze): extract imagediff subpackage, 13 -> 7 source files
The six image_diff_*.go files were a self-contained pixel-diff engine sharing a
folder with the analyze tool's JSON argument parsing. They reference nothing else in
the package -- no Deps, no parseAnalyzeArgs -- and nothing else in the package
references them, so the seam was already there; only the folder boundary was missing.
They now live in internal/tools/analyze/imagediff (5 files), leaving analyze at 7.

Five of the six files were minted by one splitting commit (fc1dc61f, "refactor: split
analyze image diff into focused modules"), so the shared filename prefix is not
independent evidence about where the seams are. Reading the call graph instead
contradicted two of those filenames:

- image_diff_math.go grouped DiffVerdict and absDiff16 as "math", but they have
  disjoint callers: DiffVerdict is called only by CompareImages and its result is the
  public DiffResult.Verdict field, while absDiff16 is called only by
  RebuildChangedGrid. "Math" was a name, not a behavior. DiffVerdict moved next to
  DiffResult, absDiff16 next to its sole caller, and the file is gone.
- image_diff_io.go held only LoadImage while image_diff_render.go did the PNG
  encode and os.Create. Both are the package's filesystem boundary, so they merged
  into imageio.go -- one file now contains every disk touch, leaving the grid,
  region and compare code pure.

Test relocation caught the per-package coverage trap: TestVisualDiffVerdict lived in
visual_diff_test.go (baseline-argument parsing) but exercises DiffVerdict, which is
image-diff code. Left behind, it would have reported DiffVerdict at 0.0% in the new
package. It moves with the code it tests. image_diff_test.go moves wholesale -- it
already depended on the unexported findChangedRegions.

Also corrects three stale names in the analyze package doc found while editing it:
it advertised DiffImages and ValidateURLs (neither exists; the real names are
CompareImages and ValidateLinksServerSide) and listed Region as an analyze type
after it moved. Documentation-only.

Behavior preservation is mechanical: the exported symbol set across the two packages
is identical (32 symbols) and the multiset of non-comment code lines is identical at
458. Combined coverage across analyze+imagediff is 74.9%, exactly the pre-split
analyze figure; imagediff is at 93.7% with no function at 0.0%. Only one consumer
changed (cmd/browser-agent/tools_analyze_visual.go), and it is an import plus four
call-site qualifiers.
